+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 ### **Copyrigth**: 
 *Tiledesk SRL*
 
+
+
+# this branch 
+- **bug-fix**: when I change the text of an ask kb action I lose the value of the selected kb
+
 # 1.39.32 
 - **bug-fix**: GptTask Preview broken
 - **bug-fix**: Condition blocks not saved when changing Boolean operators